Relocate docker folder
